domains: review cnevids
